% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ctmax.r
\name{calc.ctmax}
\alias{calc.ctmax}
\title{Calculates Cmax, Tmax from raw data for each PK curve (define using group_by)}
\usage{
calc.ctmax(x, timevar = "time", depvar = "dv")
}
\arguments{
\item{x}{a dataset}

\item{timevar}{a column name}

\item{depvar}{a column name}
}
\value{
A dataset with estimates for the Cmax (maximum concentration) and Tmax (time of first occurence of cmax) parameters, one observation per subject
}
\description{
Input dataset can contain all uncorrected data, including LOQ; estimate first occurence of maximum concentration for each PK curve; if all concentrations are NA, sets Cmax and Tmax also to NA
}
