% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.par.th.r
\name{calc.par.th}
\alias{calc.par.th}
\title{Calculates PK parameters that need lambda_z}
\usage{
calc.par.th(x = par, th = th, cov = cov, dosevar = "dose",
  factor = 1, reg = "sd", ss = "n")
}
\arguments{
\item{x}{result parameter dataset from calc.par}

\item{th}{result dataset from est.thalf}

\item{cov}{covariates dataset (containing at least dose for CL calculation)}

\item{dosevar}{variable containing the dose amount}

\item{factor}{conversion factor for CL and V calculation (e.g. dose in mg, conc in ng/mL, factor=1000)}

\item{reg}{regimen, "sd" or "md"}

\item{ss}{is steady state reached (y/n)}
}
\value{
dataset with estimates for the following parameters, one observation per subject: \cr
  all parameters calculated in th \cr
  all parameters calculated in par \cr
  clast.pred: predicted concentration at tlast \cr
  aucinf.obs: aucinf based on observed concentration at tlast \cr
  aucinf.pred: aucinf based on predicted concentration at tlast \cr
  aumcinf.obs: area under the first moment curve extrapolated to infinity, based on observed concentration at tlast \cr
  aumcinf.pred: area under the first moment curve extrapolated to infinity, based on predicted concentration at tlast \cr
  cl.f.obs: clearance based on aucinf.obs, at steady state based on auctau \cr
  cl.f.pred: clearance based on aucinf.pred \cr
  mrt.obs: Mean residence time based on aumcinf.obs and aucinf.obs \cr
  mrt.pred: Mean residence time based on aumcinf.pred and aucinf.pred \cr
  vz.f.obs: distribution volume based on cl.f.obs, at steady state based on auctau \cr
  vz.f.pred: distribution based on cl.f.pred \cr
  vss.obs: Steady-state volume based on cl.obs and mrt.obs \cr
  vss.pred: Steady-state volume based on cl.pred and mrt.pred \cr
  regimen (reg) \cr
  steady state reached Y/N? (ss) \cr
NOTE: ctmax must be merged separately as those were calculated from uncorrected data \cr
}
\description{
Calculates PK parameters that need lambda_z
}
\examples{
#'# We need half-lives for corrections, so first let's get that.
th = Theoph \%>\%
 group_by(Subject=as.numeric(Subject)) \%>\%
 do(est.thalf(.,timevar="Time",depvar="conc",includeCmax="Y")) \%>\%
 ungroup()

# We need nominal time variable as well, so let's generate that.
ID <- as.numeric(Theoph$Subject)
NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph1 <- Theoph \%>\% mutate(NTAD=metrumrg::snap(Time, NTAD))

#let's say we want AUC0-8. We only have 7 and 9 hr concentrations, so we need to interpolate conc for 8 hr.
tc = Theoph1 \%>\%
group_by(Subject=as.numeric(Subject)) \%>\%
do(correct.time(.,nomtimevar="NTAD",timevar="Time",depvar="conc",
                 tau=,tstart=,tend=,teval=8,th=th,reg="sd")) \%>\%
#above step added timepoints that we will add interpolated concentrations to with this next step
do(correct.conc(.,nomtimevar="NTAD",tau=,tstart=,tend=,teval=8,
th=th,reg="sd",ss="n")) \%>\%
ungroup()
#Now get parameters that do not require LAMBDAZ
par <- tc \%>\%
     group_by(Subject=as.numeric(Subject)) \%>\%
     ##tstart by default is 0, so here we will get AUC0-8hr and AUC0-24 as sampling ends at 24 hr
     do(calc.par(.,tau=NA, tstart=NA, tend=NA, teval=8, route="po", method=1)) \%>\%
     ungroup()
cov <- data.frame(Subject=as.numeric(Theoph1$Subject), DOSE=Theoph$Dose) \%>\%
       distinct(.,.keep_all = T)
par <- calc.par.th(x=par,th=th ,cov=cov,
                 dose="DOSE",factor=1,
                 reg="sd",ss="n")
}
