% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNCAnames.r
\name{convertNCAnames}
\alias{convertNCAnames}
\title{Format NCA parameter names for output in latex nca table}
\usage{
convertNCAnames(ds)
}
\arguments{
\item{ds}{}
}
\description{
Format NCA parameter names for output in latex nca table
}
