% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.time.r
\name{correct.time}
\alias{correct.time}
\title{Corrects concentrations at critical, but deviating time points (e.g, predose, TAU, start and end of user selected AUC interval), and adds missing records at these critical time points.}
\usage{
correct.time(x, nomtimevar = "ntad", timevar = "time", depvar = "dv",
  tau = NA, tstart = NA, tend = NA, teval = NA, th = NA,
  reg = "sd", method = 1)
}
\arguments{
\item{x}{input dataset name (contains all data, including LOQ (set conc to zero for these))}

\item{nomtimevar}{variable name containing the nominal sampling time}

\item{timevar}{variable name containing the actual sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{tau}{dosing interval (for multiple dosing), if single dose, leave empty}

\item{teval}{user selected AUC interval, if not requested, leave empty}

\item{th}{file name of file with lamdba_z information for each curve (can be derived from est.thalf)}

\item{reg}{regimen, "sd" or "md"}

\item{method}{method of interpolation: \cr
1: linear up - linear down \cr
2: linear up - logarithmic down}
}
\value{
a dataset with time deviation corrections applied (timevar and depvar adapted). The following variables are added: \cr
\tabular{cc}{
  create.nr        \tab         is a missing record created? \cr
  create.txt       \tab         explanation of what is created \cr
  trule.nr         \tab         correction rule number \cr
  trule.txt        \tab         text explaining what was altered \cr
  applies.to.time   \tab        lists all AUCS to which the time deviation rule applies \cr
  time.tau, conc.tau  \tab      time and conc, corrected for AUCtau calculation \cr
  time.teval, conc.teval \tab   time and conc, corrected for AUCteval calculation (AUC0-teval) \cr
  time.part, conc.part   \tab   time and conc, corrected for partial AUC calculation (AUCstart-end, start>0) \cr
  time.lastall, conc.lastall \tab time and conc, corrected for AUClast and AUCall calculation \cr
  t0.flag, tau.flag, tstart.flag, tend.flag, teval.flag \tab flags for what timepoint the correction was needed \cr
}
}
\description{
\itemize Records with missing NOMINAL time will be removed, this must be corrected before the function is called.
\itemize If a record at the critical time point is missing, add it and set time to nominal time and set dv conc to NA
\itemize If there is a measurable concentration AFTER the nominal time point (i.e. sample is taken too late), use interpolation
\itemize If there is NO measurable concentration AFTER the nominal time point (i.e. sample is taken too early), use extrapolation
\itemize Set deviating time at predose to 0
\itemize Original time and conc will be kept in original variables.
\itemize The following Time Deviation Correction Rules will be applied to critical time points (t=0, tau, tstart, tend, teval), if needed:\cr
\tabular{cccc}{
  Rule \tab Regimen \tab Description \tab Applied to \cr
  SDT-1 \tab sd \tab Set actual time to 0 \tab t=0 \cr
  SDT-2 \tab sd \tab Correct concentration at deviating time by interpolation \tab t=tau,tstart,tend,teval \cr
  SDT-3 \tab sd \tab Correct concentration at deviating time by extrapolation \tab t=tau,tend,teval \cr
  MDT-1 \tab md \tab If predose sample taken after dosing, set actual time to 0 and conc to NA \tab t=0 \cr
  MDT-2 \tab md \tab Correct concentration at deviating time by interpolation (too late) \tab  t=tau,tstart,tend,teval \cr
  MDT-3 \tab md \tab Correct concentration at deviating time by extrapolation (too early) \tab t=0,tau,tend,teval \cr
  MDT-3a \tab md \tab Set actual time to zero if concentration is BLOQ (too early) \tab t=0 \cr
}
}
\examples{
# We need half-lives for this, so first let's get that.
th = Theoph \%>\%
group_by(Subject=as.numeric(Subject)) \%>\%
 do(est.thalf(.,timevar="Time",depvar="conc",includeCmax="Y")) \%>\%
 ungroup()

# We need nominal time variable as well, so let's generate that.
ID <- as.numeric(Theoph$Subject)
NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph1 <- Theoph \%>\% mutate(NTAD=metrumrg::snap(Time, NTAD))

#let's say we want AUC0-8. We only have 7 and 9 hr concentrations, so we need to interpolate conc for 8 hr.
tc = Theoph1 \%>\%
group_by(Subject=as.numeric(Subject)) \%>\%
do(correct.time(.,nomtimevar="NTAD",timevar="Time",depvar="conc",
                 tau=,tstart=,tend=,teval=8,th=th,reg="sd"))
}
