% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConcTab.R
\name{formatConcTab}
\alias{formatConcTab}
\title{Format raw concentrations table to latex suitable for qP NCA reports}
\usage{
formatConcTab(ds, dsTimes = NULL, nsig = 3, omits = NULL)
}
\arguments{
\item{omits}{}
}
\description{
Format raw concentrations table to latex suitable for qP NCA reports
}
