% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTab.r
\name{formatTab}
\alias{formatTab}
\title{Format the subject-level NCA results tables and return Latex-ready tables}
\usage{
formatTab(myLst)
}
\arguments{
\item{myLst}{}
}
\description{
Format the subject-level NCA results tables and return Latex-ready tables
}
