% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol.r
\name{interpol}
\alias{interpol}
\title{Interpolate concentrations}
\usage{
interpol(c1 = NA, c2 = NA, t1 = NA, t2 = NA, t3 = NA,
  method = 1)
}
\arguments{
\item{method}{}
}
\description{
Interpolate concentrations
}
