% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.lead.r
\name{lag_lead}
\alias{lag_lead}
\title{Estimate Lagging and Leading Times and Concentrations}
\usage{
lag_lead(
  x,
  nomtimevar1 = NA,
  depvar1 = NA,
  timevar1 = NA,
  lagc = NA,
  lagt = NA,
  leadc = NA,
  leadt = NA,
  ...
)
}
\arguments{
\item{x}{data.frame}

\item{nomtimevar1}{column name in x indicating nominal time}

\item{depvar1}{column name in x indicating concentration}

\item{timevar1}{column name in x indicating time}

\item{lagc}{concentration at previous sampling time}

\item{lagt}{previous sampling time}

\item{leadc}{concentration at next sampling time}

\item{leadt}{next sampling time}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Estimates lagging and leading times and concentrations.
Used by correct.xx functions to estimate lagging and leading timepoints
and concentrations for each timepoint.
}
