% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listing.Conc.r
\name{Listing.Conc}
\alias{Listing.Conc}
\title{Wide Concentrations vs. Nominal Time Tables for NCA Report}
\usage{
Listing.Conc(ds, subjVar, timeVar, dvVar, blqVar, carryAlong = NULL, nsig = 3)
}
\arguments{
\item{ds}{data.frame}

\item{subjVar}{column name in ds indicating subject}

\item{timeVar}{column name in ds indicating time}

\item{blqVar}{column name in ds indicating concentrations below the limit of quantitation}

\item{carryAlong}{column names to carry along}

\item{nsig}{number of signifcant digits}

\item{concVar}{column name in ds indicating concentration}
}
\description{
Wide Concentrations vs. Nominal Time Tables for NCA Report
}
\examples{
library(dplyr)
library(magrittr)
NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph \%<>\%
 mutate(
 NTAD=metrumrg::snap(Time, NTAD)) \%>\%
 mutate(
   Subject=as.numeric(as.character(Subject)),
   #converting from factor to numeric
   BQL = ifelse(conc<=0.25, 1, 0),
   #just adding few BLQs to demonstrate functionality
   conc= ifelse(conc<=0.25, NA, round(conc, 2))
   #just adding digits to demonstrate nsig functionality
 )

test <- Listing.Conc(
  Theoph,
  subjVar="Subject",
  timeVar="NTAD",
  dvVar="conc",
  blqVar="BQL",
  carryAlong="Wt", nsig=3
)
}
