% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listing.Param.r
\name{Listing.Param}
\alias{Listing.Param}
\title{Generate NCA Parameter Table}
\usage{
Listing.Param(ds, subjVar, nsig = 3)
}
\arguments{
\item{ds}{data.frame}

\item{subjVar}{column name in ds indicating subject}

\item{nsig}{number of significant digits}
}
\description{
Generates NCA parameter table.
}
\examples{
library(dplyr)
# Listing_Param_theo = Listing.Param(NCAResults$pkpar, subjVar="ID", nsig=3)
}
