% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.sumstat.conc.r
\name{nca.sumstat.conc}
\alias{nca.sumstat.conc}
\title{Summarizes a vector concentration measurements for output to a report}
\usage{
nca.sumstat.conc(
  x,
  LOQ,
  ns = 3,
  bloqRule = "BLOQ=0",
  oddCode = Cs(M, NS),
  bloqCode = "BLOQ",
  ...
)
}
\arguments{
\item{x}{character vector containing numeric values to be summarized}

\item{LOQ}{limit of quantitation for x}

\item{ns}{number of significant figures for summary values}

\item{bloqRule}{Rule for treatment of BLOQ values "BLOQ=0" or "BLOQ=LOQ/2"}

\item{oddCode}{alphabetical codes for concentrations}

\item{bloqCode}{alphabetic code for BLOQ values}
}
\value{
summary statistics of concentrations vs. time
}
\description{
Summarizes a vector concentration measurements for output to a report
}
\examples{
library(dplyr)
NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph1 <- Theoph \%>\%
 mutate(NTAD=metrumrg::snap(Time, NTAD)) \%>\%
 mutate(
   Subject=as.numeric(as.character(Subject)),
   #converting from factor to numeric
   BQL = ifelse(conc<=0.25, 1, 0),
   #just adding few BLQs to demonstrate functionality
   conc= ifelse(conc<=0.25, NA, conc)
   #just adding few BLQs to demonstrate functionality
)
ConcSumTab = tapply(Theoph1$conc, Theoph1$NTAD,
                    nca.sumstat.conc,
                    LOQ = 0.250,
                    ns = 3,
                    bloqRule = "BLOQ=0",
                    oddCode = "M",
                    bloqCode = "BLOQ") \%>\%
             bind_cols()  #convert to data frame

#add names
ConcSumTab1 <- cbind(Parameter = names(ConcSumTab$`0`), ConcSumTab)

}
