% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.sumstat.params.r
\name{nca.sumstat.params}
\alias{nca.sumstat.params}
\title{Create summary statistics of NCA parameters}
\usage{
nca.sumstat.params(x, ns = 3, na.rm = T, ...)
}
\arguments{
\item{...}{}
}
\description{
Create summary statistics of NCA parameters
}
