% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaTabulate.r
\name{ncaTabulate}
\alias{ncaTabulate}
\title{Format the table NCA parameters table}
\usage{
ncaTabulate(ds, units = parUnits)
}
\arguments{
\item{ds}{a resulting data frame from nca.summary.table}

\item{units}{a data frame containing units}
}
\description{
Format the table NCA parameters table
}
