% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reg.r
\name{plot.reg}
\alias{plot.reg}
\title{A log-linear plot will be made for each curve. If elimination half-life was estimated for that curve,
the following will be indicated in the plot:}
\usage{
\method{plot}{reg}(
  x,
  by = c("id"),
  th = NA,
  bloqvar = "bloq",
  timevar = "tad",
  depvar = "dv",
  timelab = "timevar",
  deplab = "depvar",
  exclvar = NA,
  plotdir = NA
)
}
\arguments{
\item{x}{input dataset name (if called within dplyr: .)}

\item{by}{by-variable(s), e.g. c("subject","day")}

\item{th}{file name of file with half-life estimation information for each curve}

\item{bloqvar}{variable name containing the BLOQ flag (0: no, 1: yes)}

\item{timevar}{variable name containing the sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{timelab}{X-axis label (default: "timevar")}

\item{deplab}{Y-axis label (default: "depvar")}

\item{exclvar}{variable name containing information about points to be excluded (these should have <exclvar>=1)}

\item{savedir}{folder where individual plot files will be saved}
}
\value{
If the attribute 'savedir'is empty, plots will be generated in standard output, otherwise plots will be saved as
PNG file in the designated folder
PLOT.REG: Plot regression curves
}
\description{
- Cmax (Yellow, even if no half-life was estimated)
- points used in regression and resulting regression line (green)
- points excluded from regression (red crossed)
- estimate of elimination half-life and adjusted R-squared
}
\details{
Input dataset:
- uncorrected dataset, used for half-life estimation
- dataset containing results of the half-life estimation

USAGE:
plot.reg(x,by=c("id","period"),th=th,bloqvar="bloq",timevar="tad",depvar="dv",exclvar="excl_th",savedir=)
}
