% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qpNCA_output.r
\name{process_qpNCA_output}
\alias{process_qpNCA_output}
\title{Processes list output from qpNCA function and provides an output with selected varibles in a data frame.}
\usage{
process_qpNCA_output(
  data = ncaresults,
  covs = c(DOSE, FORM),
  by = ID,
  keepers = c(cmax, tmax, thalf, clast.pred, aucinf.pred, cl.f.pred, vz.f.pred)
)
}
\arguments{
\item{data}{data.frame}

\item{covs}{names of covariates (can be unquoted)}

\item{by}{names of grouping variables (can be unquoted)}

\item{keepers}{names of NCA parameters to retain (can be unquoted)}
}
\value{

}
\description{
Processes list output from qpNCA function and provides an output with selected varibles in a data frame.
}
