% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciConDataFun1.r
\name{sciConDataFun1}
\alias{sciConDataFun1}
\title{Similar to ConDataFun1 from qpToolkit, but prints very small (<0.0001) or very large (>9999) numbers in scientific notation.}
\usage{
sciConDataFun1(y, nSignif, na.rm)
}
\arguments{
\item{na.rm}{}
}
\description{
Similar to ConDataFun1 from qpToolkit, but prints very small (<0.0001) or very large (>9999) numbers in scientific notation.
}
