% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciConDataFun2.r
\name{sciConDataFun2}
\alias{sciConDataFun2}
\title{Similar to ConDataFun2 from qpToolkit, but prints very small (<0.0001) or very large (>9999) numbers in scientific notation.}
\usage{
sciConDataFun2(y, nSignif, na.rm)
}
\arguments{
\item{na.rm}{}
}
\description{
Similar to ConDataFun2 from qpToolkit, but prints very small (<0.0001) or very large (>9999) numbers in scientific notation.
}
