% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumstat.Conc.r
\name{Sumstat.Conc}
\alias{Sumstat.Conc}
\title{Creates summary statistics of concenctrations by time variable and grouping variable}
\usage{
Sumstat.Conc(ds, dvVar, timeVar, by, nsig, na.rm)
}
\arguments{
\item{ds}{data.frame}

\item{dvVar}{column name in ds indicating concentration}

\item{timeVar}{column name in ds indicating time}

\item{by}{column names in ds the interaction of which indicate data subsets}

\item{nsig}{number of significant digits}

\item{na.rm}{whether to remove NA values when computing statistics}
}
\description{
Creates summary statistics of concenctrations by time variable and grouping variable
}
\examples{
library(dplyr)
Theoph1 <- Theoph \%>\% mutate(FORM=ifelse(Dose<5, "LowDose", "HighDose"))

NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph1 <- Theoph1 \%>\%
 mutate(NTAD=metrumrg::snap(Time, NTAD)) \%>\%
 mutate(
   Subject=as.numeric(as.character(Subject)),
   #converting from factor to numeric
   BQL = ifelse(conc<=0.25, 1, 0),
   #just adding few BLQs to demonstrate functionality
   conc= ifelse(conc<=0.25, NA, round(conc, 2))
   #just adding digits to demonstrate nsig functionality
 )

test <- Sumstat.Conc(Theoph1, timeVar="NTAD", dvVar="conc", nsig=3,  na.rm=TRUE, by=NULL)

}
