% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumstat.Param.r
\name{Sumstat.Param}
\alias{Sumstat.Param}
\title{Creates summary statistics of concenctrations by time variable and grouping variable}
\usage{
Sumstat.Param(ds, subjVar, by = NULL, nsig = 3, vars_ignore = "",
  keep = "", ...)
}
\description{
Creates summary statistics of concenctrations by time variable and grouping variable
}
