% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlefun.r
\name{titlefun}
\alias{titlefun}
\title{Create title for regression plots from by-variables (in plot.reg function)}
\usage{
titlefun(df, by)
}
\arguments{
\item{df}{dataset containing concentration-time information of the current curve}

\item{by}{by-variable(s), e.g. c("subject","day")}
}
\value{

}
\description{
Create title for regression plots from by-variables (in plot.reg function)
}
