% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapm.r
\name{trapm}
\alias{trapm}
\title{Trapezoidal rules used in AUMC calculations
trapm(x=, y=, method=)
remark: if c1 or c2 is 0, always linear interpolation will be performed}
\usage{
trapm(x = NA, y = NA, method = 1)
}
\arguments{
\item{x}{variable names of x coordinates}

\item{y}{variable names of y coordinates}

\item{method}{1: all linear trapedoizal rule\cr
2: linear trap. rule up / logarithmic trap. rule down\cr
3: linear trap. rule before first Tmax, logarithmic trap. rule after first Tmax \cr}
}
\description{
Trapezoidal rules used in AUMC calculations
trapm(x=, y=, method=)
remark: if c1 or c2 is 0, always linear interpolation will be performed
}
