% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ctmax.r
\name{calc.ctmax}
\alias{calc.ctmax}
\title{Calculate Cmax and Tmax}
\usage{
calc.ctmax(x, by = NULL, timevar = "time", depvar = "dv")
}
\arguments{
\item{x}{data.frame}

\item{by}{character: column names in x indicating grouping variables; default is as.character(dplyr::groups(x))}

\item{timevar}{variable name containing the actual sampling time after dose}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}
}
\value{
A dataset with estimates for the Cmax (maximum concentration)
and Tmax (time of first occurence of cmax) parameters: one observation per subject
}
\description{
Calculates Cmax and Tmax from raw data for each PK curve defined using \code{by}. \cr
}
\details{
Input dataset can contain all uncorrected data, including LOQ;
estimate first occurence of maximum concentration for each PK curve;
if all concentrations are NA, sets Cmax and Tmax also to NA.
}
\examples{
\donttest{
library(magrittr)
library(dplyr)
data(ncx)
x <- ncx
x \%<>\% group_by(subject)
x \%<>\% correct.loq
x \%>\% calc.ctmax \%>\% head
}
}
