% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ctmax.r
\name{calc.ctmax}
\alias{calc.ctmax}
\title{Calculate Cmax and Tmax}
\usage{
calc.ctmax(x, timevar = "time", depvar = "dv")
}
\arguments{
\item{x}{data.frame}

\item{timevar}{column name in x indicating time}

\item{depvar}{column name in x indicating concentration}
}
\value{
A dataset with estimates for the Cmax (maximum concentration) and Tmax (time of first occurence of cmax) parameters: one observation per subject
}
\description{
Input dataset can contain all uncorrected data, including LOQ; estimate first occurence of maximum concentration for each PK curve; if all concentrations are NA, sets Cmax and Tmax also to NA
}
\details{
Calculates Cmax and Tmax from raw data for each PK curve.
Data subsets must be defined externally, e.g. using group_by().
}
