% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.par.r
\name{calc.par}
\alias{calc.par}
\title{Calculates AUCs, tlast, clast.obs for each PK curve (define using group_by)}
\usage{
calc.par(x, tau = NA, tstart = NA, tend = NA, teval = NA,
  route = "po", method = 1)
}
\arguments{
\item{x}{a dataset (containing all data after time/concentration deviation corrections, i.e. output from correct.time and correct.conc}

\item{tau}{dosing interval (for multiple dosing), if single dose, leave empty}

\item{tstart}{starting time of user defined interval, if not requested, leave empty}

\item{tend}{end time of user defined interval, if not requested, leave empty}

\item{teval}{user selected AUC interval, if not requested, leave empty}

\item{route}{route of drug administration ("po","iv")}

\item{method}{method of trapezoidal rule
1: linear up - linear down
2: linear up - logarithmic down
3: linear before first Tmax, logarithmic after first Tmax}
}
\value{
a dataset with estimates for the following parameters, one observation per subject:
 t0.ok     = flags if t=0 concentration could be corrected/imputes. If not, no AUCs starting at t=0 are calculated;
 tlast.ok  = flags if there is at least one measurable concentration. If not, no AUClast can be calculated;
 tlast     = time of last sample with measurable concentration;
 clast.obs = observed concentration at tlast;
 aucall    = auc calculated over all observations, including values below LOQ (which are set to 0);
 auclast   = auc calculated using all observations up to and including the last measurable concentration (clast.obs at tlast);
 aumcall   = aumc calculated over all observations, including values below LOQ (which are set to 0);
 aumclast  = aumc calculated using all observations up to and including the last measurable concentration (clast.obs at tlast);
 tau       = the dosing interval (if specified);
 calc.tau  = flags if AUCtau could be calculated;
 auctau    = auc calculated over the dosing interval, only calculated if tau is specified;
 aumctau   = aumc calculated over the dosing interval, only calculated if tau is specified;
 teval     = user selected AUC interval starting at t=0 (if specified);
 calc.teval= flags if AUCteval could be calculated;
 aucxx     = auc calculated from t=0 up to/including teval, only calculated if teval is specified (xx is substituted by teval);
 calc.part = flags if AUCpart could be calculated;
 tstart    = start time of partial AUC (if specified);
 tend      = end time of partial AUC (if specified);
 aucx_y    = partial auc from time=x up to/including time=y, where x>0, only calculated if tstart and tend are specified;
 c0        = back-extrapolated concentration at t=0 for IV bolus administration
}
\description{
Calculates AUCs, tlast, clast.obs for each PK curve (define using group_by)
}
