% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.par.th.r
\name{calc.par.th}
\alias{calc.par.th}
\title{Calculate Lambda-Z Parameters}
\usage{
calc.par.th(
  x,
  by = character(0),
  th = th,
  covariates = NA,
  dose = "dose",
  factor = 1,
  reg = "SD",
  ss = "N",
  route = "EV"
)
}
\arguments{
\item{x}{result parameter dataset from calc.par}

\item{by}{column names in x indicating grouping variables}

\item{th}{result dataset from est.thalf}

\item{covariates}{covariates dataset (containing at least dose for CL calculation); defaults to unique combinations of \code{by} and \code{dose} evaluated on \code{x}; can be character name of csv file or local object}

\item{dose}{variable containing the dose amount; default 'dose' set to 1 if not in \code{names(x)}}

\item{factor}{conversion factor for CL and V calculation (e.g. dose in mg, conc in ng/mL, factor=1000)}

\item{reg}{regimen, "sd" or "md"}

\item{ss}{is steady state reached (y/n)}

\item{route}{column name in x indicating route (EV, IVB, IVI)}
}
\value{
A dataset with estimates for the following parameters,
one observation per subject:
\itemize{
\item all parameters calculated in th
\item all parameters calculated in par
\item \strong{clast.pred} predicted concentration at tlast
\item \strong{aucinf.obs} aucinf based on observed concentration at tlast
\item \strong{aucinf.pred} aucinf based on predicted concentration at tlast
\item \strong{aumcinf.obs} area under the first moment curve extrapolated to infinity, based on observed concentration at tlast
\item \strong{aumcinf.pred} area under the first moment curve extrapolated to infinity, based on predicted concentration at tlast
\item \strong{cl.f.obs} clearance based on aucinf.obs, at steady state based on auctau
\item \strong{cl.f.pred} clearance based on aucinf.pred
\item \strong{mrt.obs} Mean residence time based on aumcinf.obs and aucinf.obs
\item \strong{mrt.pred} Mean residence time based on aumcinf.pred and aucinf.pred
\item \strong{vz.f.obs} distribution volume based on cl.f.obs, at steady state based on auctau
\item \strong{vz.f.pred} distribution based on cl.f.pred
\item \strong{vss.obs} Steady-state volume based on cl.obs and mrt.obs
\item \strong{vss.pred} Steady-state volume based on cl.pred and mrt.pred
\item \strong{reg} regimen: SD or MD
\item \strong{ss} steady state reached Y/N?
}

Note: ctmax must be merged separately as those were calculated from uncorrected data.
}
\description{
Calculates PK parameters that need lambda_z.
}
\examples{
example(calc.par) # creates x, par, th, ctmax, corrtab
# notice x includes (optional) loqrule, includeCmax, reg, method, route, ss
covs <- Theoph \%>\%
  select(subject = Subject, Wt, dose = Dose) \%>\%
  unique \%>\%
  mutate(dose = dose * Wt) # see ?Theoph
y <- x \%>\% select(subject, reg, ss, loqrule) \%>\% unique
y \%<>\% mutate(factor = 1)
par \%<>\% left_join(y, by = 'subject')
par \%<>\% calc.par.th(by = 'subject', th = th, covariates = covs)
par \%<>\% left_join(ctmax, ., by = 'subject')
par \%>\% head
par \%>\% data.frame \%>\% head(2)
}
