% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.conc.r
\name{correct.conc}
\alias{correct.conc}
\title{Corrects missing concentration at critical time points (e.g, predose, TAU, start and end of user selected AUC interval)}
\usage{
correct.conc(
  x,
  by = "subject",
  nomtimevar = "ntad",
  tau = NA,
  tstart = NA,
  tend = NA,
  teval = NA,
  th = NA,
  reg = "SD",
  ss = "N",
  route = "EV",
  method = 1
)
}
\arguments{
\item{x}{input dataset name input dataset name (contains all data, including LOQ (set conc to zero for these))}

\item{nomtimevar}{variable name containing the nominal sampling time}

\item{tau}{dosing interval (for multiple dosing), if single dose, leave empty}

\item{tstart}{start time of partial AUC (start>0), if not requested, leave empty}

\item{tend}{end time of partial AUC, if not requested, leave empty}

\item{teval}{user selected AUC interval, if not requested, leave empty}

\item{th}{file name of file with lamdba_z information for each curve (can be derived from est.thalf)}

\item{reg}{regimen, "sd" or "md"}

\item{ss}{is steady state reached (y/n)}

\item{route}{route of drug administration ("po","iv")}

\item{method}{of interpolation: \cr
1: linear up - linear down \cr
2: linear up - logarithmic down \cr}
}
\value{
a dataset with missing concentrations imputed. The following variables are added:
\tabular{cc}{
 crule.nr     \tab    correction rule number \cr
 crule.txt     \tab   text explaining what was altered \cr
 applies.to.conc  \tab  lists all AUCS to which the concentration correction rule applies \cr
}
}
\description{
\itemize{Use interpolation if there is a measurable concentration BEFORE and AFTER the missing concentration}\cr
\itemize{Use extrapolation if there is NO measurable concentration AFTER the missing concentration}\cr
\itemize{Set missing concentration at predose to 0 (SD, non-endogenous) or value at t=TAU (steady state only)}\cr
\itemize{Set missing concentration at t=TAU to value at t=0 (steady state only)}\cr
\itemize{The following Concentration Deviation Correction Rules will be applied to critical time points (t=0, tau, tstart, tend, teval), if needed:}
\tabular{llll}{
Rule \tab Regimen \tab Description \tab Applied to \cr
SDC-1  \tab   sd   \tab       Set concentration to 0 (only non-endogenous compounds)      \tab      t=0 \cr
SDC-2   \tab    sd  \tab      impute missing concentration by interpolation   \tab                  t=tau,tstart,tend,teval\cr
SDC-3   \tab    sd    \tab    impute missing concentration by extrapolation   \tab                  t=tau,tend,teval\cr
SDC-4   \tab    sd (IV)  \tab impute missing concentration by back-extrapolation   \tab             t=0\cr
MDC-1   \tab    md  \tab      impute missing concentration by existing conc at t=0 or t=tau*  \tab  t=0,tau\cr
MDC-2   \tab    md    \tab    impute missing concentration by interpolation   \tab                   t=tau,tstart,tend,teval\cr
MDC-3   \tab    md    \tab    impute missing concentration by extrapolation   \tab                  t=tau,tend,teval\cr
MDC-4   \tab    md (IV)  \tab  impute missing concentration by back-extrapolation  \tab              t=0\cr
* only if steady state has been reached\cr
}
}
