% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.loq.r
\name{correct.loq}
\alias{correct.loq}
\title{Impute Concentrations Below the Limit of Quantitation}
\usage{
correct.loq(
  x,
  nomtimevar = "ntad",
  timevar = "time",
  depvar = "dv",
  bloqvar = "bloq",
  loqvar = "loq",
  loqrule = 1
)
}
\arguments{
\item{x}{input dataset name (if called within dplyr: .) contains all uncorrected data, including LOQ}

\item{nomtimevar}{variable name containing the nominal sampling time}

\item{timevar}{variable name containing the sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{bloqvar}{variable name containing the BLOQ flag (0: no, 1: yes)}

\item{loqvar}{variable name containing the LOQ value}

\item{loqrule}{rule number to be applied to the LOQ values in the curve. x$loqrule overrides if provided}
}
\value{
A dataset with imputed BLOQ concentrations using the chosen imputation rule
}
\description{
Imputes LOQ values according to the chosen LOQ substitution rule.
Imputations will be applied to the original depvar(no new concentration
variable will be created).
}
