% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.time.r
\name{correct.time}
\alias{correct.time}
\title{Correct Concentrations for Time Deviations}
\usage{
correct.time(
  x,
  by = "subject",
  nomtimevar = "ntad",
  timevar = "time",
  depvar = "dv",
  tau = NA,
  tstart = NA,
  tend = NA,
  teval = NA,
  th = NA,
  reg = "SD",
  method = 1
)
}
\arguments{
\item{x}{input dataset name (contains all data, including LOQ (set conc to zero for these))}

\item{by}{column names in x indicating grouping variables}

\item{nomtimevar}{variable name containing the nominal sampling time}

\item{timevar}{variable name containing the actual sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{tau}{dosing interval (for multiple dosing), if single dose, leave empty}

\item{tstart}{column name in x indicating start time}

\item{tend}{column name in x indicating end time}

\item{teval}{user selected AUC interval, if not requested, leave empty}

\item{th}{file name of file with lamdba_z information for each curve (can be derived from est.thalf)}

\item{reg}{regimen, "sd" or "md"}

\item{method}{method of interpolation:
\itemize{
\item 1: linear up - linear down
\item 2: linear up - logarithmic down
}}
}
\value{
a dataset with time deviation corrections applied (timevar and depvar adapted). The following variables are added:
\strong{Variable} | \strong{Description}
---- | -----
create.nr        |         is a missing record created?
create.txt       |         explanation of what is created
trule.nr         |         correction rule number
trule.txt        |         text explaining what was altered
applies.to.time   |        lists all AUCS to which the time deviation rule applies
time.tau, conc.tau  |      time and conc, corrected for AUCtau calculation
time.teval, conc.teval |   time and conc, corrected for AUCteval calculation (AUC0-teval)
time.part, conc.part   |   time and conc, corrected for partial AUC calculation (AUCstart-end, start>0)
time.lastall, conc.lastall | time and conc, corrected for AUClast and AUCall calculation
t0.flag, tau.flag, tstart.flag, tend.flag, teval.flag | flags for what timepoint the correction was needed
}
\description{
Corrects concentrations at critical, but deviating time points
(e.g, predose, TAU, start and end of user selected AUC interval),
and adds missing records at these critical time points.
\itemize{
\item Records with missing NOMINAL time will be removed and this must be corrected before the function is called.
\item If a record at the critical time point is missing and add it and set time to nominal time and set dv conc to NA
\item Use interpolation if there is a measurable concentration AFTER the nominal time point (i.e. sample is taken too late)
\item Use extrapilation if there is NO measurable concentration AFTER the nominal time point (i.e. sample is taken too early)
\item Set deviating time at predose to 0
\item Original time and conc will be kept in original variables.
\item The following Time Deviation Correction Rules will be applied to critical time points (t=0, tau, tstart, tend, teval), if needed:
}
}
\details{

}
