% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.thalf.r
\name{est.thalf}
\alias{est.thalf}
\title{Calculates lambda_z and thalf for each PK curve (defined using group_by)}
\usage{
est.thalf(x, timevar = "time", depvar = "dv", includeCmax = "Y")
}
\arguments{
\item{x}{a dataset (not needed to be corrected time and conc)}

\item{timevar}{variable name containing the sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{includeCmax}{include results of regression including Cmax in selection? (y/n)}
}
\value{
a dataset with estimates for each regression analysis in one observation. The following parameters are available:
no.points = number of data points used in the regression analysis; \cr
intercept = estimated intercept;\cr
lambda_z = -1*estimated slope;\cr
r.squared = square of the correlation coefficient;\cr
adj.r.squared = adjusted square of the correlation coefficient;\cr
thalf = elimination half-life;\cr
start_th = time of first sample included in the thalf estimation;\cr
end_th = time of last sample included in the thalf estimation;\cr
includeCmax = include results of regression including Cmax in selection? (y/n)\cr
}
\description{
The function starts with the last three sample points and performs log-linear regression on it. It then adds one sampling point at a time (including and ending at tmax) and performs the regression again. Internal variable EST checks whether there are at least 3 timepoints for estimation after removal of LOQs and NAs
}
\examples{
th = Theoph \%>\%
 group_by(Subject=as.numeric(Subject)) \%>\%
 do(est.thalf(.,timevar="Time",depvar="conc",includeCmax="Y")) \%>\%
 ungroup()
}
