% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenamefun.r
\name{filenamefun}
\alias{filenamefun}
\title{Create file name for regression plots (*.png) from by-variables (in plot.reg function)}
\usage{
filenamefun(df, by)
}
\arguments{
\item{by}{}
}
\description{
Create file name for regression plots (*.png) from by-variables (in plot.reg function)
}
