% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reg.r
\name{filenamefun}
\alias{filenamefun}
\title{Create File Name for Regression Plots.
Creates file name for regression plots (*.png) from by-variables in plot_reg function.
Also used by nca.sum().}
\usage{
filenamefun(x, by)
}
\arguments{
\item{x}{data.frame}

\item{by}{column names in x indicating grouping variables}
}
\value{
character
}
\description{
Create File Name for Regression Plots.
Creates file name for regression plots (*.png) from by-variables in plot_reg function.
Also used by nca.sum().
}
