% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.Table.r
\name{Format.Table}
\alias{Format.Table}
\title{Format the table NCA parameters table}
\usage{
Format.Table(ds, units = parUnits)
}
\arguments{
\item{ds}{a resulting data frame from nca.summary.table}

\item{units}{a data frame containing units}
}
\description{
Format the table NCA parameters table
}
