% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.lead.r
\name{lag.lead}
\alias{lag.lead}
\title{Estimate lagging and leading time points and concentrations for each time point}
\usage{
lag.lead(x, nomtimevar1 = NA, depvar1 = NA, timevar1 = NA,
  lagc = NA, lagt = NA, leadc = NA, leadt = NA)
}
\arguments{
\item{leadt}{}
}
\description{
Estimate lagging and leading time points and concentrations for each time point
}
