% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listing.Sumstat.Param.r
\name{Listing.Sumstat.Param}
\alias{Listing.Sumstat.Param}
\title{Creates a table of individual parameterss table with summary statistics}
\usage{
Listing.Sumstat.Param(ds, subjVar, by, nsig, vars_ignore = "", keep = "", ...)
}
\arguments{
\item{ds}{Dataframe of NCA results}

\item{subjVar}{Name of individual identifier variable}

\item{by}{Grouping variable}

\item{nsig}{Number of significant digits}

\item{vars_ignore}{List of variables that need to be ignored for summary statistics based on a filtering condition}

\item{keep}{List of variables that you want to keep in addition throughout the summary (such as the filtering and grouping variables.}
}
\description{
Creates a table of individual parameterss table with summary statistics
}
