% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.conc.table.r
\name{nca.conc.table}
\alias{nca.conc.table}
\title{Format concentrations summary tables for report ready}
\usage{
nca.conc.table(obs, LOQ = 1e-04, sumVar = "conc", subjVar = "subjid",
  timeVar = "nom.time", omits = NULL, times = NULL,
  carryAlong = NULL, nsig = 3, bloqCode = "BLOQ")
}
\arguments{
\item{bloqCode}{}
}
\description{
Format concentrations summary tables for report ready
}
