% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.format.table.r
\name{nca.format.table}
\alias{nca.format.table}
\title{Format NCA Parameters Table}
\usage{
nca.format.table(
  nca,
  carryAlong = NULL,
  parms = Cs(Cmax, Tmax, AUCall, AUCINF.pred, Cl.F.pred, HL.Lambda.z),
  ns = rep(3, length(parms)),
  uAdjust = rep(1, length(parms))
)
}
\arguments{
\item{nca}{parameters table}

\item{carryAlong}{column names to carry along}

\item{parms}{parameter names to format}

\item{ns}{number of significant digits}

\item{uAdjust}{factor by which to adjust all parameters}
}
\value{

}
\description{
Formats an NCA parameters table.
}
