% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.sum.r
\name{nca.sum}
\alias{nca.sum}
\title{Create summarising PDF for each curve}
\usage{
nca.sum(x, corrfile = corrtab, by = c("subject"), plotdir = NA, pdfdir = NA)
}
\arguments{
\item{pdfdir}{}
}
\value{

}
\description{
Create summarising PDF for each curve
}
