% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.sum.r
\name{nca.sum}
\alias{nca.sum}
\title{Create Summary for Each Curve}
\usage{
nca.sum(x, by = character(0), corrfile, pdfdir = NA)
}
\arguments{
\item{x}{data.frame}

\item{by}{column names in x indicating grouping variables}

\item{corrfile}{a correction table like the output of \code{\link{tab.corr}}}

\item{pdfdir}{path to directory for pdf ouput}
}
\value{
(invisible) pdfdir
}
\description{
Creates a PDF summary for each curve.
}
\examples{
example(qpNCA)
}
