% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.summary.table.r
\name{nca.summary.table}
\alias{nca.summary.table}
\title{Summarize NCA parameters table}
\usage{
nca.summary.table(
  ds,
  subjVar = "subjid",
  carryAlong = NULL,
  omits = NULL,
  nsig = rep(3, ncol(ds) - 1 - length(carryAlong)),
  LOQ = 1e-04
)
}
\arguments{
\item{ds}{formatted nca parameter table (output from nca.format.table)}

\item{subjVar}{variable name of subject identifier}

\item{carryAlong}{character vector of variable names to be included in output table}

\item{omits}{vector of subject identifiers (subjid) to be omitted from calculation of summary statistics}

\item{nsig}{vector of default number of significant figures to report parameter values}

\item{LOQ}{limit of quantitation for concentration variables}
}
\value{

}
\description{
Summarize NCA parameters table
}
