% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.sumstat.conc.r
\name{nca.sumstat.conc}
\alias{nca.sumstat.conc}
\title{Generate summary statistics of concentrations}
\usage{
nca.sumstat.conc(x, LOQ, na.rm = T, ns = 3, bloqRule = "BLOQ=0",
  oddCode = Cs(M, NS), bloqCode = "BLOQ", ...)
}
\arguments{
\item{...}{}
}
\description{
Generate summary statistics of concentrations
}
