% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpNCA.r
\name{qpNCA}
\alias{qpNCA}
\title{Perform Non-compartmental Analysis}
\usage{
qpNCA(
  x,
  by = c("subject"),
  nomtimevar = "ntad",
  timevar = "time",
  depvar = "dv",
  bloqvar = "bloq",
  loqvar = "loq",
  loqrule = 1,
  includeCmax = "Y",
  exclvar = NA,
  plotdir = NA,
  pdfdir = NA,
  timelab = "timevar",
  deplab = "depvar",
  tau = NA,
  tstart = NA,
  tend = NA,
  teval = NA,
  covfile = NA,
  dose = NA,
  factor = NA,
  reg = "SD",
  ss = "N",
  route = "EV",
  method = 1
)
}
\arguments{
\item{x}{input dataset name (if called within dplyr: .)}

\item{by}{by-variable(s), e.g. c("subject","day")}

\item{nomtimevar}{variable name containing the nominal sampling time}

\item{timevar}{variable name containing the sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{bloqvar}{variable name containing the BLOQ flag (0: no, 1: yes)}

\item{loqvar}{variable name containing the LOQ value}

\item{loqrule}{rule number to be applied to the LOQ values in the curve; x$loqrule overrides if provided}

\item{includeCmax}{include results of regression including Cmax in selection? (y/n) x$includeCmax overrides if provided}

\item{exclvar}{variable name containing information about points to be excluded (these should have <exclvar>=1)}

\item{plotdir}{folder where regression plots (.PNG) will be saved; leave empty for standard output}

\item{tau}{dosing interval (for multiple dosing), if single dose, leave empty; x$tau overrides if provided}

\item{tstart}{start time of partial AUC (start>0), if not requested, leave empty; x$tstart overrides if provided}

\item{tend}{end time of partial AUC, if not requested, leave empty; x$tend overrides if provided}

\item{teval}{user selected AUC interval, if not requested, leave empty; x$tend overrides if provided}

\item{dose}{variable containing the dose amount}

\item{factor}{conversion factor for CL and V calculation (e.g. dose in mg, conc in ng/mL, factor=1000); x$factor overrides if provided}

\item{reg}{regimen, "sd" or "md"; can be character column name in x; x$regimen overrides if provided}

\item{ss}{is steady state reached (y/n); x$ss overrides if provided}

\item{route}{route of drug administration ("EV","IVB", "IVI"); x$overrides if provided}

\item{method}{method for trapezoidal rule;  x$method overrides if provided
1: linear up - linear down
2: linear up - logarithmic down
3: linear before first Tmax, logarithmic after first Tmax}

\item{X-axis}{label (default: "timevar") in plots}

\item{Y-axis}{label (default: "depvar") in plots}

\item{cov}{covariates dataset}
}
\value{
covariates : contains covariates selected with the 'cov' argument
half_life  : contains linear regression parameters
ctmax      : contains cmax and tmax estimated from uncorrected data
ct_corr    : contains the time and concentration corrected dataset
corrections: contains descriptions of the corrections applied
pkpar      : contains all estimated PK parameters
}
\description{
Consecutively executes the following NCA steps:
\describe{
 \item{\code{link{correct.loq}}}{impute LOQ values}
 \item{\code{link{est.thalf}}}{calculate lambda-z and halflife}
 \item{\code{link{plot.reg}}}{plot each curve}
 \item{\code{link{calc.ctmax}}}{calculate Cmax and Tmax}
 \item{\code{link{correct.time}}}{supply records at critical times with correct concentration}
 \item{\code{link{correct.conc}}}{supply correct concentrations at critical times}
 \item{\code{link{tab.corr}}}{tabulate data alterations}
 \item{\code{link{calc.par}}}{calculates profile-specific summary statistics}
 \item{\code{link{calc.par.th}}}{calculates parameters depednent on lambda-z}
 }
}
\details{

}
