% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpNCA.r
\name{qpNCA}
\alias{qpNCA}
\title{A wrapper for the individual qpNCA functions
This wrapper function consecutively executes the following qpNCA functions:
1. correct.loq
2. est.thalf
2a. plot.reg
3. calc.ctmax
4. correct.time
5. correct.conc
6. tab.corr
7. calc.par
8. calc.par.th}
\usage{
qpNCA(x, by = c("subject"), nomtimevar = "ntad", timevar = "time",
  depvar = "dv", bloqvar = "bloq", loqvar = "loq", loqrule = 1,
  includeCmax = "Y", excl = "excl", savedir = NA,
  timelab = "timevar", deplab = "depvar", tau = NA, tstart = NA,
  tend = NA, teval = NA, cov = NA, dose = NA, factor = NA,
  reg = "SD", ss = "N", route = "PO", method = 1)
}
\arguments{
\item{x}{input dataset name (if called within dplyr: .)}

\item{by}{by-variable(s), e.g. c("subject","day")}

\item{nomtimevar}{variable name containing the nominal sampling time}

\item{timevar}{variable name containing the sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{bloqvar}{variable name containing the BLOQ flag (0: no, 1: yes)}

\item{loqvar}{variable name containing the LOQ value}

\item{loqrule}{rule number to be applied to the LOQ values in the curve}

\item{includeCmax}{include results of regression including Cmax in selection? (y/n)}

\item{excl}{variable name containing information about points to be excluded (these should have <excl>=1)}

\item{savedir}{folder where regression plots (.PNG) will be saved; leave empty for standard output}

\item{tau}{dosing interval (for multiple dosing), if single dose, leave empty}

\item{tstart}{start time of partial AUC (start>0), if not requested, leave empty}

\item{tend}{end time of partial AUC, if not requested, leave empty}

\item{teval}{user selected AUC interval, if not requested, leave empty}

\item{cov}{covariates dataset}

\item{dose}{variable containing the dose amount}

\item{factor}{onversion factor for CL and V calculation (e.g. dose in mg, conc in ng/mL, factor=1000)}

\item{reg}{regimen, "sd" or "md"}

\item{ss}{is steady state reached (y/n)}

\item{route}{route of drug administration ("po","iv")}

\item{method}{method for trapezoidal rule:
1: linear up - linear down
2: linear up - logarithmic down
3: linear before first Tmax, logarithmic after first Tmax}

\item{X-axis}{label (default: "timevar") in plots}

\item{Y-axis}{label (default: "depvar") in plots}
}
\value{
covariates : contains covariates selected with the 'cov' argument
half_life  : contains linear regression parameters
ctmax      : contains cmax and tmax estimated from uncorrected data
ct_corr    : contains the time and concentration corrected dataset
corrections: contains descriptions of the corrections applied
pkpar      : contains all estimated PK parameters
}
\description{
A wrapper for the individual qpNCA functions
This wrapper function consecutively executes the following qpNCA functions:
1. correct.loq
2. est.thalf
2a. plot.reg
3. calc.ctmax
4. correct.time
5. correct.conc
6. tab.corr
7. calc.par
8. calc.par.th
}
