% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpNCA.r
\name{qpNCA}
\alias{qpNCA}
\title{Perform Non-compartmental Analysis}
\usage{
qpNCA(
  x,
  by = character(0),
  nomtimevar = "ntad",
  timevar = "time",
  depvar = "dv",
  bloqvar = "bloq",
  loqvar = "loq",
  loqrule = 1,
  includeCmax = "Y",
  exclvar = NA,
  plotdir = NA,
  pdfdir = NA,
  timelab = "timevar",
  deplab = "depvar",
  tau = NA,
  tstart = NA,
  tend = NA,
  teval = NA,
  covariates = NA,
  dose = NA,
  factor = NA,
  reg = "SD",
  ss = "N",
  route = "EV",
  method = 1
)
}
\arguments{
\item{x}{input dataset name}

\item{by}{column names in x indicating grouping variables}

\item{nomtimevar}{variable name containing the nominal sampling time}

\item{timevar}{variable name containing the sampling time}

\item{depvar}{variable name containing the dependent variable (e.g., concentration)}

\item{bloqvar}{variable name containing the BLOQ flag (0: no, 1: yes)}

\item{loqvar}{variable name containing the LOQ value}

\item{loqrule}{rule number to be applied to the LOQ values in the curve; x$loqrule overrides if provided}

\item{includeCmax}{include results of regression including Cmax in selection? (y/n) x$includeCmax overrides if provided}

\item{exclvar}{variable name containing information about points to be excluded (these should have exclvar = 1)}

\item{plotdir}{directory where regression plots (.PNG) will be saved; NA gives default location, NULL skips regression plots}

\item{pdfdir}{directory where pdf summaries will be saved; NA gives default location, NULL skips summary}

\item{timelab}{label for time axis}

\item{deplab}{label for dependent variable axis}

\item{tau}{dosing interval (for multiple dosing); NA (default) for if single dose; x$tau overrides}

\item{tstart}{start time of partial AUC (start>0); NA (default) if not requested; x$tstart overrides}

\item{tend}{end time of partial AUC; NA (default) if not requested; x$tend overrides}

\item{teval}{user selected AUC interval; NA (default) if not requested; x$teval overrides}

\item{covariates}{covariates dataset}

\item{dose}{variable containing the dose amount}

\item{factor}{conversion factor for CL and V calculation (e.g. dose in mg, conc in ng/mL, factor=1000); x$factor overrides if provided}

\item{reg}{regimen, "sd" or "md"; can be character column name in x; x$regimen overrides if provided}

\item{ss}{is steady state reached (y/n); x$ss overrides if provided}

\item{route}{route of drug administration ("EV","IVB", "IVI"); x$overrides if provided}

\item{method}{method for trapezoidal rule;  x$method overrides if provided
\itemize{
\item 1: linear up - linear down
\item 2: linear up - logarithmic down
\item 3: linear before first Tmax, logarithmic after first Tmax
}}
}
\value{
(list)
\itemize{
\item \strong{covariates} covariates selected with the \code{cov} argument
\item \strong{half_life} linear regression parameters
\item \strong{ctmax} cmax and tmax estimated from uncorrected data
\item \strong{ct_corr} the time and concentration corrected dataset
\item \strong{corrections} descriptions of the corrections applied
\item \strong{pkpar} all estimated PK parameters
\item \strong{plotdir} directory for plots
\item \strong{pdfdir} directory for pdf summaries
}
}
\description{
Consecutively executes the following NCA steps:
\itemize{
\item \link{correct.loq} impute LOQ values
\item \link{est.thalf} calculate lambda-z and halflife
\item \link{plot_reg} plot each curve
\item \link{calc.ctmax} calculate Cmax and Tmax
\item \link{correct.time} supply records at critical times with correct concentration
\item \link{correct.conc} supply correct concentrations at critical times
\item \link{tab.corr} tabulate data alterations
\item \link{calc.par} calculates profile-specific summary statistics
\item \link{calc.par.th} calculates parameters dependent on lambda-z
}
}
