% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumstat.Param.r
\name{Sumstat.Param}
\alias{Sumstat.Param}
\title{Creates summary statistics of concenctrations by time variable and grouping variable}
\usage{
Sumstat.Param(ds, subjVar, by = NULL, nsig = 3, vars_ignore = "",
  vars_keep = "", ...)
}
\arguments{
\item{ds}{Dataframe of NCA results}

\item{subjVar}{Name of individual identifier variable}

\item{by}{Grouping variable}

\item{nsig}{Number of significant digits}

\item{vars_ignore}{List of variables that need to be ignored for summary statistics based on a filtering condition}

\item{keep}{List of variables that you want to keep in addition throughout the summary (such as the filtering and grouping variables.}
}
\description{
Creates summary statistics of concenctrations by time variable and grouping variable
}
