% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reg.r
\name{titlefun}
\alias{titlefun}
\title{Create Title for Regression Plots}
\usage{
titlefun(x, by)
}
\arguments{
\item{x}{dataset containing concentration-time information of the current curve}

\item{by}{column names in x indicating grouping variables}
}
\value{
character
}
\description{
Creates title for regression plots in plot_reg() using by-variables.
}
