% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap.r
\name{trap}
\alias{trap}
\title{Calculate Area Under the Curve Using Trapezoids}
\usage{
trap(x = NA, y = NA, method = 1)
}
\arguments{
\item{x}{x variable, i.e. time}

\item{y}{y variable, i.e. concentration}

\item{method}{method:
\itemize{
\item 1: linear up - linear down
\item 2: linear up - logarithmic down
\item 3: linear before Tmax, logarithmic after Tmax
}}
}
\value{
area (length-one numeric)
}
\description{
Calculates AUC using the trapezoidal method. Assumes data represent a single profile.
Despite choice of method, only linear interpolation is used
for areas of intervals beginning or ending with y: 0.
}
