% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapm.r
\name{trapm}
\alias{trapm}
\title{Calculate Area Under the Moment Curve Using Trapezoids}
\usage{
trapm(x = NA, y = NA, method = 1)
}
\arguments{
\item{x}{variable names of x coordinates}

\item{y}{variable names of y coordinates}

\item{method}{method:
\itemize{
\item 1: linear up - linear down
\item 2: linear up - logarithmic down
\item 3: linear before Tmax, logarithmic after Tmax
}}
}
\value{
area (length-one numeric)
}
\description{
Calculates AUMC using the trapezoidal method.
Assumes data represent a single profile.
Despite choice of method, only linear interpolation is used
for areas of intervals beginning or ending with y: 0.
}
